changed AssignRoom to Reservations, Reservations to Hotel, updated tests- for better design
