+++ specs/spec_helper.rb
@@ -6,6 +6,6 @@ require 'minitest/reporters'
 require 'minitest/skip_dsl'
 require_relative '../lib/room.rb'
 require_relative '../lib/reservation.rb'
-require_relative '../lib/assign_room.rb'
+require_relative '../lib/hotel.rb'
 
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
